addressed comments

+++ mem0/memory/main.py
@@ -35,6 +35,7 @@ class Memory(MemoryBase):
         self.llm = LlmFactory.create(self.config.llm.provider, self.config.llm.config)
         self.db = SQLiteManager(self.config.history_db_path)
         self.collection_name = self.config.vector_store.config.collection_name
+        self.version = self.config.version
 
         self.enable_graph = False
         if self.config.graph_store.provider:
@@ -164,7 +165,7 @@ class Memory(MemoryBase):
                 )
         capture_event("mem0.add", self)
 
-        if self.enable_graph:
+        if self.version == "v1.1" and self.enable_graph:
             if user_id:
                 self.graph.user_id = user_id
             else:
@@ -243,19 +244,22 @@ class Memory(MemoryBase):
             }
             for mem in memories[0]
         ]
-
-        if self.enable_graph:
-            graph_entities = self.graph.get_all()
-            return {"memories": all_memories, "entities": graph_entities}
-
-        warnings.warn(
-            "Call to deprecated function get_all. "
-            "Returning a list of memories is deprecated. "
-            "Use the new dict format {'memories': []} instead.",
-            category=DeprecationWarning,
-            stacklevel=2
-        )
-        return all_memories
+        
+        if self.version == "v1.1":
+            if self.enable_graph:
+                graph_entities = self.graph.get_all()
+                return {"memories": all_memories, "entities": graph_entities}
+            else:
+                return {"memories" : all_memories}
+        else:
+            warnings.warn(
+                "The current get_all API output format is deprecated. "
+                "To use the latest format, set `api_version='v1.1'`. "
+                "The current format will be removed in mem0ai 1.1.0 and later versions.",
+                category=DeprecationWarning,
+                stacklevel=2
+            )
+            return all_memories
     
 
     def search(
@@ -283,7 +287,7 @@ class Memory(MemoryBase):
         if run_id:
             filters["run_id"] = run_id
 
-        capture_event("mem0.search", self, {"filters": len(filters), "limit": limit})
+        capture_event("mem0.search", self, {"filters": len(filters), "limit": limit, "version": self.version})
         embeddings = self.embedding_model.embed(query)
         memories = self.vector_store.search(query=embeddings, limit=limit, filters=filters)
 
@@ -306,18 +310,21 @@ class Memory(MemoryBase):
             for mem in memories
         ]
 
-        if self.enable_graph:
-            graph_entities = self.graph.search(query)
-            return {"memories": original_memories, "entities": graph_entities}
-
-        warnings.warn(
-            "Call to deprecated function search. "
-            "Returning a list of memories is deprecated. "
-            "Use the new dict format {'memories': []} instead.",
-            category=DeprecationWarning,
-            stacklevel=2
-        )
-        return original_memories
+        if self.version == "v1.1":
+            if self.enable_graph:
+                graph_entities = self.graph.search(query)
+                return {"memories": original_memories, "entities": graph_entities}
+            else:
+                return {"memories" : original_memories}
+        else:
+            warnings.warn(
+                "The current get_all API output format is deprecated. "
+                "To use the latest format, set `api_version='v1.1'`. "
+                "The current format will be removed in mem0ai 1.1.0 and later versions.",
+                category=DeprecationWarning,
+                stacklevel=2
+            )
+            return original_memories
 
     def update(self, memory_id, data):
         """
@@ -372,7 +379,7 @@ class Memory(MemoryBase):
         for memory in memories:
             self._delete_memory_tool(memory.id)
 
-        if self.enable_graph:
+        if self.version == "v1.1" and self.enable_graph:
             self.graph.delete_all()
 
         return {'message': 'Memories deleted successfully!'}
